+++ app/chitter.rb
@@ -2,10 +2,12 @@ require './app/data_mapper_setup'
 
 require './app/controllers/base'
 require './app/controllers/user'
+require './app/controllers/post'
 
 
 module Application
   class Chitter < Sinatra::Base
-    use Routes::XUser
+    use Routes::Users
+    use Routes::Posts
   end
 end
