+++ app/controllers/post.rb
@@ -1,0 +1,13 @@
+require_relative 'base'
+
+module Application
+  module Routes
+    class Posts < Base
+
+      get '/posts/new' do
+        erb :'/posts/new'
+      end
+
+    end
+  end
+end
