+++ spec/helpers/session.rb
@@ -1,0 +1,11 @@
+#require 'spec_helper'
+#module SessionHelpers
+  def sign_up user
+    visit '/users/new'
+    fill_in :email, with: user.email
+    fill_in :name, with: user.name
+    fill_in :username, with: user.username
+    fill_in :password, with: user.password
+    click_button 'Sign Up'
+  end
+#end
