+++ spec/spec_helper.rb
@@ -1,9 +1,11 @@
 # Generated by rspec-sinatra. (2015-08-15 13:12:27 +0100)
 
+
 ENV['RACK_ENV'] = 'test'
 
 require File.join(File.dirname(__FILE__), '..', 'app/chitter.rb')
 
+#require 'helpers/session'
 require './app/data_mapper_setup'
 require 'capybara'
 require 'capybara/rspec'
